usage hint about UninstallNotLoadedAddons

+++ autoload/scriptmanager.vim
@@ -173,6 +173,10 @@ endf
 
 fun! scriptmanager#UninstallAddons(list)
   let list = a:list
+  if list == []
+    echo "no pluigns selected. If you ran UninstallNotLoadedAddons use <tab> or <c-d> to get a list of not loaded plugins"
+    return
+  endif
   call map(list, 'scriptmanager#PluginDirByName(v:val)')
   if input('confirm running rm -fr on plugins:'.join(list,",").' [y/n]') == 'y'
     for path in list
